Fix #1852: Use the new_username method from srusers.
This ensures that all user-creation follows the same rules for new
user names, since the underlying library now contains the logic in
use.

Note that the actual fix for this ticket is in the logic in the
library method -- this commit just changes this client to use that
method instead of the local version.

Change-Id: Ia305070ced7819e53e45c1e0d1f46ebb30a39d07