Add scripts to help create kickstart users.
getteamsdata extracts the team data from trac, outputting it to a
file (teams_map.csv) for storage and inspection.

kickstartusers creates a user per team from the teams_map.csv file,
ensuring that this user is a student, and is in the right team and
college groups (which it creates if needed).

Change-Id: I69390b6a0a59f2e2eea66fbd17738471fe21c8b9